Split out computeHugoniot and computeHugoniotWithPorosity as functions available and installed with Spheral so we can use them with any EOS. The planar compaction problem uses these utilities now to compute the analytic solution.
